# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/Admin/Logs/Activities/LoadData": {
            "post": {
                "tags": [
                    "Activity Logs"
                ],
                "summary": "Get all activity logs (List)",
                "description": "Retrieve paginated list of all admin activity logs with search and sorting capabilities",
                "operationId": "93af731e2e72af7dd2ec5474bbb10f95",
                "requestBody": {
                    "description": "Optional parameters for filtering, pagination, and sorting",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "description": "Global search by event, description, or admin name",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "pageNumber": {
                                        "description": "Page number",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "pageSize": {
                                        "description": "Items per page",
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "sortColumn": {
                                        "description": "Column to sort by (id, event, description, created_at)",
                                        "type": "string",
                                        "example": "created_at"
                                    },
                                    "sortColumnDir": {
                                        "description": "Sort direction (asc or desc)",
                                        "type": "string",
                                        "enum": [
                                            "asc",
                                            "desc"
                                        ],
                                        "example": "desc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "no": {
                                                        "description": "Nomor urut",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "event": {
                                                        "type": "string",
                                                        "example": "login"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "User logged in"
                                                    },
                                                    "severity": {
                                                        "type": "string",
                                                        "enum": [
                                                            "info",
                                                            "warning",
                                                            "error",
                                                            "success"
                                                        ],
                                                        "example": "success"
                                                    },
                                                    "severityColor": {
                                                        "type": "string",
                                                        "example": "#10B981"
                                                    },
                                                    "severityIcon": {
                                                        "type": "string",
                                                        "example": "✅"
                                                    },
                                                    "severityLabel": {
                                                        "type": "string",
                                                        "example": "Success"
                                                    },
                                                    "admin": {
                                                        "$ref": "#/components/schemas/AdminResource"
                                                    },
                                                    "createdAt": {
                                                        "type": "string",
                                                        "example": "2025-01-15 10:30:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "pageSize": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "totalPages": {
                                                    "type": "integer",
                                                    "example": 7
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/Admin/Logs/Activities/View/{id}": {
            "post": {
                "tags": [
                    "Activity Logs"
                ],
                "summary": "View single activity log",
                "description": "Retrieve detailed information about a specific activity log",
                "operationId": "6a64199af5dd67094de6208449b9aa09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Activity Log ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "event": {
                                                    "type": "string",
                                                    "example": "login"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "User logged in"
                                                },
                                                "severity": {
                                                    "type": "string",
                                                    "enum": [
                                                        "info",
                                                        "warning",
                                                        "error",
                                                        "success"
                                                    ],
                                                    "example": "success"
                                                },
                                                "severityColor": {
                                                    "type": "string",
                                                    "example": "#10B981"
                                                },
                                                "severityIcon": {
                                                    "type": "string",
                                                    "example": "✅"
                                                },
                                                "severityLabel": {
                                                    "type": "string",
                                                    "example": "Success"
                                                },
                                                "admin": {
                                                    "$ref": "#/components/schemas/AdminResource"
                                                },
                                                "properties": {
                                                    "type": "object",
                                                    "example": {
                                                        "old": {
                                                            "name": "John"
                                                        },
                                                        "new": {
                                                            "name": "Jane"
                                                        }
                                                    }
                                                },
                                                "createdAt": {
                                                    "type": "string",
                                                    "example": "2025-01-15 10:30:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/Admin/Logs/Activities/Dashboard": {
            "post": {
                "tags": [
                    "Activity Logs"
                ],
                "summary": "Get activity dashboard statistics",
                "description": "Retrieve activity statistics for dashboard including recent activities, summary, top events, and most active users",
                "operationId": "8a3869ab4f6f1b9627b2a480ec411675",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "recent_activities": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "activity_summary": {
                                                    "properties": {
                                                        "today": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "this_week": {
                                                            "type": "integer",
                                                            "example": 120
                                                        },
                                                        "this_month": {
                                                            "type": "integer",
                                                            "example": 450
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "top_events": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "most_active_users": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminResource": {
                "title": "Admin Resource",
                "description": "Schema for the admin resource",
                "properties": {
                    "no": {
                        "type": "integer",
                        "example": 1
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "firstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "fullName": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "Super Admin"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "supplierId": {
                        "type": "string",
                        "format": "uuid",
                        "example": null,
                        "nullable": true
                    },
                    "supplierCode": {
                        "type": "string",
                        "example": "SUP-001",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "isSuperAdmin": {
                        "type": "boolean",
                        "example": true
                    },
                    "createdAt": {
                        "type": "string",
                        "example": "2025-01-15 10:30:00"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "401": {
                "description": "Forbidden: Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated"
                                },
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "statusCode": {
                                    "type": "integer",
                                    "example": 401
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "404": {
                "description": "Resource Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Resource Not Found"
                                },
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "statusCode": {
                                    "type": "integer",
                                    "example": 404
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}